ref(ngrx-ducks): extract interfaces into own files
