+++ ngrx/ducks/src/lib/typings/dispatcher-for-effect.ts
@@ -1,0 +1,3 @@
+export type DispatcherForEffect = {
+  type: string;
+};
