+++ ngrx/ducks/src/lib/typings/effect-dispatcher.ts
@@ -1,0 +1,17 @@
+export type EffectDispatcher<T> = T extends {
+  type: string;
+  action: () => void;
+}
+  ? {
+      type: string;
+      dispatch: () => void;
+    }
+  : T extends {
+      type: string;
+      action: (payload: infer TPayload) => void;
+    }
+  ? {
+      type: string;
+      dispatch: (payload: TPayload) => void;
+    }
+  : never;
